+++ plugin/src/main/groovy/org/openbakery/AbstractXcodeBuildTask.groovy
@@ -116,7 +116,7 @@ abstract class AbstractXcodeBuildTask extends DefaultTask {
 		}
 
 		if (project.xcodebuild.signing.keychainPathInternal.exists()) {
-			commandList.add('OTHER_CODE_SIGN_FLAGS="--keychain ' + project.xcodebuild.signing.keychainPathInternal.path + '"');
+			commandList.add('OTHER_CODE_SIGN_FLAGS=--keychain ' + project.xcodebuild.signing.keychainPathInternal.path);
 		}
 
 		return commandList;