+++ plugin/src/main/groovy/org/openbakery/HockeyAppPluginExtension.groovy
@@ -21,6 +21,7 @@ import org.gradle.api.Project
 class HockeyAppPluginExtension {
 	def Object outputDirectory
 	def String apiToken = null
+	def String appID = null
 	def String notes = "This build was uploaded using the gradle xcodePlugin"
 	def String status = 2
 	def String notify = 1